+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/preview-middleware
 
+## 0.17.29
+
+### Patch Changes
+
+-   063a2f4: fix: enable manifest actions for v2 apps with array page structure
+
+## 0.17.28
+
+### Patch Changes
+
+-   d95bade: fix: Fixed various bugs related to Enable Variant Management for Tables quick action. It was unnecessarily disabled in some apps on List Report; changing Object Page table type led to enabling this action again; action is disabled now for custom tables, where it can't be applied.
+
+## 0.17.27
+
+### Patch Changes
+
+-   7c76e7c: fix: unify contextual menu for all nodes in outline panel.
+
 ## 0.17.26
 
 ### Patch Changes